ice skill tree and areas to be privated later
made the skills of the ice skill tree (4 of them) but i couldnt do the tree itself. the abilites work fine other than the ice breath ability and the last normal ability. the boolean variables that tell if the player has chosen a certain path is public and interactable for testing + dmeo purposes